+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### New
+
+-   Custom field values now display in payment details (#5647)
+
 ### Changed
 
 -   Fields API throws an exception when inserting relative to a node that does not exist (#5640)
+-   Form Templates now support custom checkbox fields. (#5643)
 -   Custom fields now support arbitrary attributes (#5641)
 
 ## 2.9.6 - 2021-01-13